+++ src/natgw.js
@@ -1,0 +1,72 @@
+const { PUBLIC_SUBNET } = require('./constants');
+
+/**
+ * Build an EIP
+ *
+ * @param {Number} position
+ * @return {Object}
+ */
+function buildEIP(position) {
+  const cfName = `EIP${position}`;
+  return {
+    [cfName]: {
+      Type: 'AWS::EC2::EIP',
+      Properties: {
+        Domain: 'vpc',
+      },
+    },
+  };
+}
+
+/**
+ * Build a NatGateway in a given AZ
+ *
+ * @param {Number} position
+ * @param {String} zone
+ * @param {String} stage
+ * @return {Object}
+ */
+function buildNatGateway(position, zone, stage) {
+  const cfName = `NatGateway${position}`;
+  return {
+    [cfName]: {
+      Type: 'AWS::EC2::NatGateway',
+      Properties: {
+        AllocationId: {
+          'Fn::GetAtt': [
+            `EIP${position}`,
+            'AllocationId',
+          ],
+        },
+        SubnetId: {
+          Ref: `${PUBLIC_SUBNET}Subnet${position}`,
+        },
+        Tags: [
+          {
+            Key: 'STAGE',
+            Value: stage,
+          },
+          {
+            Key: 'Name',
+            Value: {
+              'Fn::Join': [
+                '-',
+                [
+                  {
+                    Ref: 'AWS::StackName',
+                  },
+                  zone,
+                ],
+              ],
+            },
+          },
+        ],
+      },
+    },
+  };
+}
+
+module.exports = {
+  buildEIP,
+  buildNatGateway,
+};
